requirements.txt
alembic==1.0.3
blinker==1.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Admin==1.5.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
markdown2==2.3.6
MarkupSafe==1.1.0
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
